chore: consolidate context directory
