+++ backend/src/contexts/AuthContext.js
@@ -1,49 +1,45 @@
-// src/context/AuthContext.js
+// src/contexts/AuthContext.js
 import React, { createContext, useState, useContext, useEffect } from 'react';
 
 const AuthContext = createContext(null);
 
 export const AuthProvider = ({ children }) => {
-  const [authState, setAuthState] = useState(() => {
-    // console.log('[AuthContext] useState 초기화 함수 실행');
+  const getInitialAuthState = () => {
     const token = localStorage.getItem('token');
     const userString = localStorage.getItem('user');
-    // console.log('[AuthContext] 초기 localStorage - token:', token, 'userString:', userString);
-
     if (token && userString) {
       try {
         const user = JSON.parse(userString);
-        // console.log('[AuthContext] 초기 상태: 로컬 스토리지 유효, isAuthenticated: true, user:', user);
         return { token, user, isAuthenticated: true };
       } catch (error) {
-        console.error('[AuthContext] 초기 상태: 로컬 스토리지 user 파싱 오류:', error);
+        // 파싱 오류 시 저장된 정보 제거
         localStorage.removeItem('token');
         localStorage.removeItem('user');
         return { token: null, user: null, isAuthenticated: false };
       }
     }
-    // console.log('[AuthContext] 초기 상태: 로컬 스토리지 비어있음 또는 정보 부족, isAuthenticated: false');
     return { token: null, user: null, isAuthenticated: false };
-  });
+  };
 
-  useEffect(() => {
-    // console.log('[AuthContext] authState 변경 감지:', authState);
-  }, [authState]);
+  const [authState, setAuthState] = useState(getInitialAuthState);
+
+  // useEffect는 로컬 스토리지 변경을 감지하는 용도보다는,
+  // 외부 요인(예: 다른 탭에서의 로그아웃)에 의한 동기화에 더 적합할 수 있습니다.
+  // 여기서는 초기 상태 설정을 getInitialAuthState 함수로 명확히 했습니다.
+  // 만약 다른 탭과의 동기화가 필요하다면 'storage' 이벤트를 리스닝할 수 있습니다.
 
   const login = (token, user) => {
     localStorage.setItem('token', token);
     localStorage.setItem('user', JSON.stringify(user));
-    const newState = { token, user, isAuthenticated: true };
-    setAuthState(newState);
-    console.log('[AuthContext] login 함수 호출됨, 새로운 authState:', newState);
+    setAuthState({ token, user, isAuthenticated: true });
+    console.log('로그인 상태 업데이트 (AuthContext):', { user });
   };
 
   const logout = () => {
     localStorage.removeItem('token');
     localStorage.removeItem('user');
-    const newState = { token: null, user: null, isAuthenticated: false };
-    setAuthState(newState);
-    console.log('[AuthContext] logout 함수 호출됨, 새로운 authState:', newState);
+    setAuthState({ token: null, user: null, isAuthenticated: false });
+    console.log('로그아웃 완료 (AuthContext)');
   };
 
   const value = { authState, login, logout };
@@ -58,7 +54,7 @@ export const AuthProvider = ({ children }) => {
 export const useAuth = () => {
   const context = useContext(AuthContext);
   if (context === undefined) {
-    throw new Error('useAuth는 AuthProvider 내부에서 사용해야 합니다.');
+    throw new Error('useAuth must be used within an AuthProvider');
   }
   return context;
 };
