+++ src/components/common/Header.js
@@ -1,7 +1,7 @@
 // src/components/common/Header.js
 import React from 'react';
 import { Link, useNavigate } from 'react-router-dom';
-import { useAuth } from '../../context/AuthContext';
+import { useAuth } from '../../contexts/AuthContext';
 import './Header.css';
 
 const Header = () => {
