--- src/context/ThemeContext.js
@@ -1,29 +1,0 @@
-// src/context/ThemeContext.js
-import React, { createContext, useState, useContext, useMemo } from 'react';
-
-const ThemeContext = createContext(undefined);
-
-export const ThemeProvider = ({ children }) => {
-  const [theme, setTheme] = useState('light'); // 기본 테마를 'light'로 설정
-
-  const toggleTheme = () => {
-    setTheme(prevTheme => (prevTheme === 'light' ? 'dark' : 'light'));
-  };
-
-  // 테마가 변경될 때마다 body 클래스를 업데이트 (선택 사항)
-  React.useEffect(() => {
-    document.body.className = theme;
-  }, [theme]);
-
-  const value = useMemo(() => ({ theme, toggleTheme }), [theme]);
-
-  return <ThemeContext.Provider value={value}>{children}</ThemeContext.Provider>;
-};
-
-export const useTheme = () => {
-  const context = useContext(ThemeContext);
-  if (context === undefined) {
-    throw new Error('useTheme must be used within a ThemeProvider');
-  }
-  return context;
-};
