+++ src/index.js
@@ -4,10 +4,10 @@ import ReactDOM from 'react-dom';
 import './index.css';      // 전역 CSS 임포트
 import './App.css';        // App 전역 스타일
 import App from './App';
-import { ThemeProvider } from './context/ThemeContext'; // 생성된 ThemeContext 사용
-import { AuthProvider } from './context/AuthContext';   // 기존 AuthContext 사용
-import { MarketProvider } from './context/MarketContext'; // 주인님의 MarketContext 사용
-import { OrderProvider } from './context/OrderContext';   // 기존 OrderContext 사용
+import { ThemeProvider } from './contexts/ThemeContext'; // 생성된 ThemeContext 사용
+import { AuthProvider } from './contexts/AuthContext';   // 기존 AuthContext 사용
+import { MarketProvider } from './contexts/MarketContext'; // 주인님의 MarketContext 사용
+import { OrderProvider } from './contexts/OrderContext';   // 기존 OrderContext 사용
 
 // import './i18n'; // 다국어 지원 설정이 있다면 주석 해제
